Renamed project, command to xmandump
Closes #2.
